Move some dependencies into async
