+++ packages/app-store/apps.keys-schemas.generated.ts
@@ -25,6 +25,7 @@ import { appKeysSchema as vital_zod_ts } from "./vital/zod";
 import { appKeysSchema as wordpress_zod_ts } from "./wordpress/zod";
 import { appKeysSchema as zapier_zod_ts } from "./zapier/zod";
 import { appKeysSchema as zoomvideo_zod_ts } from "./zoomvideo/zod";
+
 export const appKeysSchemas = {
   dailyvideo: dailyvideo_zod_ts,
   fathom: fathom_zod_ts,
